Feat: Sync data of server
